+++ src/pmh/estimators/__init__.py
@@ -1,6 +1,6 @@
 """Lemma D1--D7 estimators for Sigma_task."""
 
-from pmh.estimators.d1_subspace import estimate_d1
+from pmh.estimators.d1_subspace import estimate_d1, estimate_d1_gram_unlabeled
 from pmh.estimators.d2_isotropic import estimate_d2
 from pmh.estimators.d3_augmentation import estimate_d3
 from pmh.estimators.d4_domain import estimate_d4
@@ -10,6 +10,7 @@ from pmh.estimators.d7_alignment import estimate_d7
 
 __all__ = [
     "estimate_d1",
+    "estimate_d1_gram_unlabeled",
     "estimate_d2",
     "estimate_d3",
     "estimate_d4",